refactor(frontend): split vendor invoice dialog state from VendorDocuments
