+++ packages/frontend/src/sections/VendorDocuments.tsx
@@ -1,10 +1,4 @@
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { api, apiResponse } from '../api';
 import { AnnotationsCard } from '../components/AnnotationsCard';
 import {
@@ -30,10 +24,10 @@ import { VendorDocumentsVendorInvoicesSection } from './vendor-documents/VendorD
 import { VendorInvoiceLineDialog } from './vendor-documents/VendorInvoiceLineDialog';
 import { VendorInvoicePoLinkDialog } from './vendor-documents/VendorInvoicePoLinkDialog';
 import { VendorInvoiceSavedViewBar } from './vendor-documents/VendorInvoiceSavedViewBar';
-import { buildVendorInvoiceLinePayload } from './vendor-documents/vendorInvoiceLinePayload';
 import { useVendorInvoiceSavedViews } from './vendor-documents/useVendorInvoiceSavedViews';
 import { useVendorDocumentsLookups } from './vendor-documents/useVendorDocumentsLookups';
 import { useVendorDocumentsTableData } from './vendor-documents/useVendorDocumentsTableData';
+import { useVendorInvoiceDialogs } from './vendor-documents/useVendorInvoiceDialogs';
 import {
   defaultPurchaseOrderForm,
   defaultVendorInvoiceForm,
@@ -54,12 +48,8 @@ import type {
   PurchaseOrder,
   PurchaseOrderDetail,
   PurchaseOrderForm,
-  PurchaseOrderLine,
   VendorInvoice,
-  VendorInvoiceAllocation,
   VendorInvoiceForm,
-  VendorInvoiceLine,
-  VendorInvoicePoLineUsage,
   VendorOption,
   VendorQuote,
   VendorQuoteForm,
@@ -130,33 +120,7 @@ export const VendorDocuments: React.FC = () => {
   const [invoicePoLinkBusy, setInvoicePoLinkBusy] = useState(false);
   const [invoicePoLinkResult, setInvoicePoLinkResult] =
     useState<MessageState>(null);
-  const [invoiceAllocationDialog, setInvoiceAllocationDialog] = useState<{
-    invoice: VendorInvoice;
-  } | null>(null);
-  const [invoiceAllocations, setInvoiceAllocations] = useState<
-    VendorInvoiceAllocation[]
-  >([]);
-  const [invoiceAllocationLoading, setInvoiceAllocationLoading] =
-    useState(false);
-  const [invoiceAllocationSaving, setInvoiceAllocationSaving] = useState(false);
-  const [invoiceAllocationMessage, setInvoiceAllocationMessage] =
-    useState<MessageState>(null);
-  const [invoiceAllocationReason, setInvoiceAllocationReason] = useState('');
-  const [invoiceAllocationExpanded, setInvoiceAllocationExpanded] =
-    useState(false);
-  const [invoiceLineDialog, setInvoiceLineDialog] = useState<{
-    invoice: VendorInvoice;
-  } | null>(null);
-  const [invoiceLines, setInvoiceLines] = useState<VendorInvoiceLine[]>([]);
-  const [invoiceLineLoading, setInvoiceLineLoading] = useState(false);
   const invoiceSavedViews = useVendorInvoiceSavedViews();
-  const [invoiceLineSaving, setInvoiceLineSaving] = useState(false);
-  const [invoiceLineMessage, setInvoiceLineMessage] =
-    useState<MessageState>(null);
-  const [invoiceLineReason, setInvoiceLineReason] = useState('');
-  const [invoiceLineExpanded, setInvoiceLineExpanded] = useState(false);
-  const [invoiceLinePoUsageByPoLineId, setInvoiceLinePoUsageByPoLineId] =
-    useState<Record<string, VendorInvoicePoLineUsage>>({});
   const [invoiceSubmitBusy, setInvoiceSubmitBusy] = useState<
     Record<string, boolean>
   >({});
@@ -172,12 +136,6 @@ export const VendorDocuments: React.FC = () => {
     id: string;
     numberLabel: string;
   } | null>(null);
-  const invoiceLineTempIdRef = useRef(0);
-
-  const nextInvoiceLineTempId = useCallback(() => {
-    invoiceLineTempIdRef.current += 1;
-    return `tmp-line-${invoiceLineTempIdRef.current}`;
-  }, []);
   const {
     availablePurchaseOrders,
     availablePurchaseOrdersForInvoicePoLink,
@@ -304,71 +262,6 @@ export const VendorDocuments: React.FC = () => {
     }
   }, []);
 
-  const loadVendorInvoiceAllocations = useCallback(
-    async (invoiceId: string) => {
-      setInvoiceAllocationLoading(true);
-      try {
-        const res = await api<{
-          invoice: VendorInvoice;
-          items: VendorInvoiceAllocation[];
-        }>(`/vendor-invoices/${invoiceId}/allocations`);
-        setInvoiceAllocations(res.items || []);
-        setInvoiceAllocationDialog((prev) =>
-          prev ? { ...prev, invoice: res.invoice } : prev,
-        );
-      } catch (err) {
-        console.error('Failed to load vendor invoice allocations.', err);
-        setInvoiceAllocationMessage({
-          text: '配賦明細の取得に失敗しました',
-          type: 'error',
-        });
-        setInvoiceAllocations([]);
-      } finally {
-        setInvoiceAllocationLoading(false);
-      }
-    },
-    [],
-  );
-
-  const loadVendorInvoiceLines = useCallback(
-    async (invoiceId: string) => {
-      setInvoiceLineLoading(true);
-      try {
-        const res = await api<{
-          invoice: VendorInvoice;
-          items: VendorInvoiceLine[];
-          poLineUsage?: VendorInvoicePoLineUsage[];
-        }>(`/vendor-invoices/${invoiceId}/lines`);
-        setInvoiceLines(
-          (res.items || []).map((item) =>
-            item.id || item.tempId
-              ? item
-              : { ...item, tempId: nextInvoiceLineTempId() },
-          ),
-        );
-        const usageByPoLineId: Record<string, VendorInvoicePoLineUsage> = {};
-        (res.poLineUsage || []).forEach((entry) => {
-          usageByPoLineId[entry.purchaseOrderLineId] = entry;
-        });
-        setInvoiceLinePoUsageByPoLineId(usageByPoLineId);
-        setInvoiceLineDialog((prev) =>
-          prev ? { ...prev, invoice: res.invoice } : prev,
-        );
-      } catch (err) {
-        console.error('Failed to load vendor invoice lines.', err);
-        setInvoiceLineMessage({
-          text: '請求明細の取得に失敗しました',
-          type: 'error',
-        });
-        setInvoiceLines([]);
-        setInvoiceLinePoUsageByPoLineId({});
-      } finally {
-        setInvoiceLineLoading(false);
-      }
-    },
-    [nextInvoiceLineTempId],
-  );
-
   useEffect(() => {
     const loadAll = async () => {
       await Promise.all([loadProjects(), loadVendors()]);
@@ -434,82 +327,51 @@ export const VendorDocuments: React.FC = () => {
     status !== 'received' && status !== 'draft' && status !== 'rejected';
   const normalizeCurrency = (value: string) =>
     value.trim().toUpperCase().slice(0, 3);
-
-  const allocationTotals = useMemo(() => {
-    if (!invoiceAllocationDialog || invoiceAllocations.length === 0)
-      return null;
-    let amountTotal = 0;
-    let taxTotal = 0;
-    invoiceAllocations.forEach((item) => {
-      amountTotal += parseNumberValue(item.amount) ?? 0;
-      taxTotal += parseNumberValue(item.taxAmount) ?? 0;
-    });
-    const grossTotal = amountTotal + taxTotal;
-    const invoiceTotal = parseNumberValue(
-      invoiceAllocationDialog.invoice.totalAmount,
-    );
-    const diff = invoiceTotal != null ? invoiceTotal - grossTotal : null;
-    return { amountTotal, taxTotal, grossTotal, invoiceTotal, diff };
-  }, [invoiceAllocationDialog, invoiceAllocations]);
-
-  const allocationTaxRateSummary = useMemo(() => {
-    const summary = new Map<string, { amount: number; tax: number }>();
-    invoiceAllocations.forEach((item) => {
-      const rateValue = parseNumberValue(item.taxRate);
-      const key = rateValue == null ? '免税' : `${rateValue}%`;
-      const entry = summary.get(key) || { amount: 0, tax: 0 };
-      entry.amount += parseNumberValue(item.amount) ?? 0;
-      entry.tax += parseNumberValue(item.taxAmount) ?? 0;
-      summary.set(key, entry);
-    });
-    return Array.from(summary.entries()).map(([key, value]) => ({
-      key,
-      amount: value.amount,
-      tax: value.tax,
-    }));
-  }, [invoiceAllocations]);
-
-  const invoiceLineTotals = useMemo(() => {
-    if (!invoiceLineDialog || invoiceLines.length === 0) return null;
-    let amountTotal = 0;
-    let taxTotal = 0;
-    let grossTotal = 0;
-    invoiceLines.forEach((line) => {
-      const quantity = parseNumberValue(line.quantity) ?? 0;
-      const unitPrice = parseNumberValue(line.unitPrice) ?? 0;
-      const amount =
-        parseNumberValue(line.amount) ?? Math.round(quantity * unitPrice);
-      const taxRate = parseNumberValue(line.taxRate);
-      const taxAmount =
-        parseNumberValue(line.taxAmount) ??
-        (taxRate == null ? 0 : Math.round((amount * taxRate) / 100));
-      amountTotal += amount;
-      taxTotal += taxAmount;
-      grossTotal += amount + taxAmount;
-    });
-    const invoiceTotal = parseNumberValue(
-      invoiceLineDialog.invoice.totalAmount,
-    );
-    const diff = invoiceTotal != null ? invoiceTotal - grossTotal : null;
-    return {
-      amountTotal,
-      taxTotal,
-      grossTotal,
-      invoiceTotal,
-      diff,
-    };
-  }, [invoiceLineDialog, invoiceLines]);
-
-  const invoiceLineRequestedQuantityByPoLine = useMemo(() => {
-    const map = new Map<string, number>();
-    invoiceLines.forEach((line) => {
-      const lineId = line.purchaseOrderLineId?.trim();
-      if (!lineId) return;
-      const quantity = parseNumberValue(line.quantity) ?? 0;
-      map.set(lineId, (map.get(lineId) || 0) + quantity);
-    });
-    return map;
-  }, [invoiceLines]);
+  const {
+    invoiceAllocationDialog,
+    invoiceAllocations,
+    invoiceAllocationLoading,
+    invoiceAllocationSaving,
+    invoiceAllocationMessage,
+    invoiceAllocationReason,
+    invoiceAllocationExpanded,
+    invoiceLineDialog,
+    invoiceLines,
+    invoiceLineLoading,
+    invoiceLineSaving,
+    invoiceLineMessage,
+    invoiceLineReason,
+    invoiceLineExpanded,
+    invoiceLinePoUsageByPoLineId,
+    invoiceLinePurchaseOrderDetail,
+    allocationTotals,
+    allocationTaxRateSummary,
+    invoiceLineTotals,
+    invoiceLineRequestedQuantityByPoLine,
+    openVendorInvoiceAllocationDialog,
+    closeVendorInvoiceAllocationDialog,
+    openVendorInvoiceLineDialog,
+    closeVendorInvoiceLineDialog,
+    addVendorInvoiceAllocationRow,
+    updateVendorInvoiceAllocation,
+    removeVendorInvoiceAllocation,
+    saveVendorInvoiceAllocations,
+    setInvoiceAllocationReason,
+    toggleInvoiceAllocationExpanded,
+    addVendorInvoiceLineRow,
+    updateVendorInvoiceLine,
+    removeVendorInvoiceLine,
+    saveVendorInvoiceLines,
+    setInvoiceLineReason,
+    toggleInvoiceLineExpanded,
+  } = useVendorInvoiceDialogs({
+    projects,
+    purchaseOrderDetails,
+    loadPurchaseOrderDetail,
+    loadVendorInvoices,
+    isVendorInvoiceAllocationReasonRequiredStatus,
+    isVendorInvoiceLineReasonRequiredStatus,
+  });
 
   const createPurchaseOrder = async () => {
     if (!poForm.projectId || !poForm.vendorId) {
@@ -819,266 +681,6 @@ export const VendorDocuments: React.FC = () => {
       });
     } finally {
       setInvoicePoLinkBusy(false);
-    }
-  };
-
-  const openVendorInvoiceAllocationDialog = async (invoice: VendorInvoice) => {
-    setInvoiceAllocationDialog({ invoice });
-    setInvoiceAllocationReason('');
-    setInvoiceAllocationMessage(null);
-    setInvoiceAllocationExpanded(false);
-    setInvoiceAllocations([]);
-    if (invoice.purchaseOrderId) {
-      void loadPurchaseOrderDetail(invoice.purchaseOrderId);
-    }
-    await loadVendorInvoiceAllocations(invoice.id);
-  };
-
-  const openVendorInvoiceLineDialog = async (invoice: VendorInvoice) => {
-    setInvoiceLineDialog({ invoice });
-    setInvoiceLineReason('');
-    setInvoiceLineMessage(null);
-    setInvoiceLineExpanded(false);
-    setInvoiceLines([]);
-    setInvoiceLinePoUsageByPoLineId({});
-    if (invoice.purchaseOrderId) {
-      void loadPurchaseOrderDetail(invoice.purchaseOrderId);
-    }
-    await loadVendorInvoiceLines(invoice.id);
-  };
-
-  const addVendorInvoiceAllocationRow = () => {
-    const defaultProjectId =
-      invoiceAllocationDialog?.invoice.projectId || projects[0]?.id || '';
-    setInvoiceAllocations((prev) => [
-      ...prev,
-      {
-        projectId: defaultProjectId,
-        amount: 0,
-        taxRate: null,
-        taxAmount: null,
-        purchaseOrderLineId: '',
-      },
-    ]);
-  };
-
-  const updateVendorInvoiceAllocation = (
-    index: number,
-    update: Partial<VendorInvoiceAllocation>,
-  ) => {
-    setInvoiceAllocations((prev) =>
-      prev.map((item, idx) => (idx === index ? { ...item, ...update } : item)),
-    );
-  };
-
-  const removeVendorInvoiceAllocation = (index: number) => {
-    setInvoiceAllocations((prev) => prev.filter((_, idx) => idx !== index));
-  };
-
-  const addVendorInvoiceLineRow = () => {
-    const maxLineNo = invoiceLines.reduce((maxValue, line) => {
-      const value = parseNumberValue(line.lineNo);
-      if (value == null || !Number.isInteger(value)) return maxValue;
-      return Math.max(maxValue, value);
-    }, 0);
-    const nextLineNo = maxLineNo + 1;
-    setInvoiceLines((prev) => [
-      ...prev,
-      {
-        tempId: nextInvoiceLineTempId(),
-        lineNo: nextLineNo,
-        description: '',
-        quantity: 1,
-        unitPrice: 0,
-        amount: null,
-        taxRate: null,
-        taxAmount: null,
-        purchaseOrderLineId: '',
-      },
-    ]);
-  };
-
-  const updateVendorInvoiceLine = (
-    index: number,
-    update: Partial<VendorInvoiceLine>,
-  ) => {
-    setInvoiceLines((prev) =>
-      prev.map((item, idx) => (idx === index ? { ...item, ...update } : item)),
-    );
-  };
-
-  const removeVendorInvoiceLine = (index: number) => {
-    setInvoiceLines((prev) => prev.filter((_, idx) => idx !== index));
-  };
-
-  const saveVendorInvoiceLines = async () => {
-    if (!invoiceLineDialog) return;
-    const invoice = invoiceLineDialog.invoice;
-    const reasonText = invoiceLineReason.trim();
-    if (
-      isVendorInvoiceLineReasonRequiredStatus(invoice.status) &&
-      !reasonText
-    ) {
-      setInvoiceLineMessage({
-        text: '変更理由を入力してください',
-        type: 'error',
-      });
-      return;
-    }
-    const payloadResult = buildVendorInvoiceLinePayload(
-      invoiceLines,
-      reasonText,
-    );
-    if (!payloadResult.ok) {
-      setInvoiceLineMessage({ text: payloadResult.errorText, type: 'error' });
-      return;
-    }
-
-    try {
-      setInvoiceLineSaving(true);
-      setInvoiceLineMessage(null);
-      await api(`/vendor-invoices/${invoice.id}/lines`, {
-        method: 'PUT',
-        body: JSON.stringify(payloadResult.payload),
-      });
-      setInvoiceLineMessage({
-        text: '請求明細を更新しました',
-        type: 'success',
-      });
-      await loadVendorInvoices();
-      await loadVendorInvoiceLines(invoice.id);
-    } catch (err) {
-      console.error('Failed to update vendor invoice lines.', err);
-      const errorText = err instanceof Error ? err.message : String(err);
-      if (errorText.includes('PO_LINE_QUANTITY_EXCEEDED')) {
-        setInvoiceLineMessage({
-          text: 'PO明細の数量上限を超えています（数量を見直してください）',
-          type: 'error',
-        });
-      } else if (errorText.includes('LINE_TOTAL_MISMATCH')) {
-        setInvoiceLineMessage({
-          text: '請求合計との差分が解消されていません',
-          type: 'error',
-        });
-      } else if (errorText.includes('INVALID_PURCHASE_ORDER_LINE')) {
-        setInvoiceLineMessage({
-          text: '選択したPO明細が関連POに属していません',
-          type: 'error',
-        });
-      } else {
-        setInvoiceLineMessage({
-          text: '請求明細の更新に失敗しました',
-          type: 'error',
-        });
-      }
-    } finally {
-      setInvoiceLineSaving(false);
-    }
-  };
-
-  const saveVendorInvoiceAllocations = async () => {
-    if (!invoiceAllocationDialog) return;
-    const invoice = invoiceAllocationDialog.invoice;
-    const reasonText = invoiceAllocationReason.trim();
-    if (
-      isVendorInvoiceAllocationReasonRequiredStatus(invoice.status) &&
-      !reasonText
-    ) {
-      setInvoiceAllocationMessage({
-        text: '変更理由を入力してください',
-        type: 'error',
-      });
-      return;
-    }
-    const payload: {
-      allocations: Array<{
-        projectId: string;
-        amount: number;
-        taxRate?: number;
-        taxAmount?: number;
-        purchaseOrderLineId?: string;
-      }>;
-      reasonText?: string;
-    } = { allocations: [] };
-
-    for (let i = 0; i < invoiceAllocations.length; i += 1) {
-      const entry = invoiceAllocations[i];
-      const projectId = entry.projectId.trim();
-      if (!projectId) {
-        setInvoiceAllocationMessage({
-          text: `配賦明細 ${i + 1} の案件が未選択です`,
-          type: 'error',
-        });
-        return;
-      }
-      const amount = parseNumberValue(entry.amount);
-      if (amount == null || amount < 0) {
-        setInvoiceAllocationMessage({
-          text: `配賦明細 ${i + 1} の金額が不正です`,
-          type: 'error',
-        });
-        return;
-      }
-      const taxRate =
-        entry.taxRate === undefined ||
-        entry.taxRate === null ||
-        entry.taxRate === ''
-          ? null
-          : parseNumberValue(entry.taxRate);
-      if (entry.taxRate != null && taxRate == null) {
-        setInvoiceAllocationMessage({
-          text: `配賦明細 ${i + 1} の税率が不正です`,
-          type: 'error',
-        });
-        return;
-      }
-      const taxAmount =
-        entry.taxAmount === undefined ||
-        entry.taxAmount === null ||
-        entry.taxAmount === ''
-          ? null
-          : parseNumberValue(entry.taxAmount);
-      if (entry.taxAmount != null && taxAmount == null) {
-        setInvoiceAllocationMessage({
-          text: `配賦明細 ${i + 1} の税額が不正です`,
-          type: 'error',
-        });
-        return;
-      }
-      const purchaseOrderLineId = entry.purchaseOrderLineId?.trim();
-      payload.allocations.push({
-        projectId,
-        amount,
-        ...(taxRate != null ? { taxRate } : {}),
-        ...(taxAmount != null ? { taxAmount } : {}),
-        ...(purchaseOrderLineId ? { purchaseOrderLineId } : {}),
-      });
-    }
-    if (reasonText) {
-      payload.reasonText = reasonText;
-    }
-
-    try {
-      setInvoiceAllocationSaving(true);
-      setInvoiceAllocationMessage(null);
-      await api(`/vendor-invoices/${invoice.id}/allocations`, {
-        method: 'PUT',
-        body: JSON.stringify(payload),
-      });
-      setInvoiceAllocationMessage({
-        text: '配賦明細を更新しました',
-        type: 'success',
-      });
-      await loadVendorInvoices();
-      await loadVendorInvoiceAllocations(invoice.id);
-    } catch (err) {
-      console.error('Failed to update vendor invoice allocations.', err);
-      setInvoiceAllocationMessage({
-        text: '配賦明細の更新に失敗しました',
-        type: 'error',
-      });
-    } finally {
-      setInvoiceAllocationSaving(false);
     }
   };
 
@@ -1404,10 +1006,6 @@ export const VendorDocuments: React.FC = () => {
   const activePo = poSendLogDialogId
     ? purchaseOrderMap.get(poSendLogDialogId)
     : null;
-  const invoiceLinePurchaseOrderDetail = invoiceLineDialog?.invoice
-    .purchaseOrderId
-    ? purchaseOrderDetails[invoiceLineDialog.invoice.purchaseOrderId] || null
-    : null;
 
   return (
     <div>
@@ -1614,9 +1212,9 @@ export const VendorDocuments: React.FC = () => {
         allocationTaxRateSummary={allocationTaxRateSummary}
         reason={invoiceAllocationReason}
         message={invoiceAllocationMessage}
-        onClose={() => setInvoiceAllocationDialog(null)}
+        onClose={closeVendorInvoiceAllocationDialog}
         onSave={saveVendorInvoiceAllocations}
-        onToggleExpanded={() => setInvoiceAllocationExpanded((prev) => !prev)}
+        onToggleExpanded={toggleInvoiceAllocationExpanded}
         onAddRow={addVendorInvoiceAllocationRow}
         onUpdateAllocation={updateVendorInvoiceAllocation}
         onRemoveAllocation={removeVendorInvoiceAllocation}
@@ -1644,15 +1242,15 @@ export const VendorDocuments: React.FC = () => {
         reason={invoiceLineReason}
         message={invoiceLineMessage}
         missingNumberLabel={missingNumberLabel}
-        onClose={() => setInvoiceLineDialog(null)}
+        onClose={closeVendorInvoiceLineDialog}
         onSave={saveVendorInvoiceLines}
-        onToggleExpanded={() => setInvoiceLineExpanded((prev) => !prev)}
+        onToggleExpanded={toggleInvoiceLineExpanded}
         onAddRow={addVendorInvoiceLineRow}
         onUpdateLine={updateVendorInvoiceLine}
         onRemoveLine={removeVendorInvoiceLine}
         onChangeReason={setInvoiceLineReason}
         onOpenAllocation={(invoice) => {
-          setInvoiceLineDialog(null);
+          closeVendorInvoiceLineDialog();
           void openVendorInvoiceAllocationDialog(invoice);
         }}
         renderProject={renderProject}
